+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Renamed `customArgv` to `argv`, which requires removing first two arguments on node
 - Replaced `args` in options with `arg`, which passes by `--example=arg`
 - Now if no `cliName` passed, it sets `unnamed-cli`
+- Updated return value of args from string to object
 
 ## v2.1.0 2024-07-18